Remove test definition
Tests can be arbitrarily named so it is difficult to search for them
appropriately. Further, they tend to pollute the search results of all
other items and just get in the way.

There are benefits of course to being able to search for tests, but
these are outweighed by the drawbacks right now in my opinion.